+++ R/mat2tab.R
@@ -4,7 +4,7 @@
 #'
 #' @param mat : \code{matrix}
 #'
-#' @return tab : \code{data.frame} with 2 integer (IDs) and 1 numeric (values) columns.
+#' @return \code{data.frame} with 2 integer (IDs) and 1 numeric (values) columns.
 #'
 #' @examples
 #' ped = data.frame(ID=1:6, SIRE=c(0,0,1,3,1,4), DAM=c(0,0,2,2,2,5))