+++ R/smgsped.R
@@ -6,7 +6,7 @@
 #'
 #' @param pheno : Vector of phenotyped individuals
 #'
-#' @return newped : S-MGS pedigree \code{data.frame}
+#' @return S-MGS pedigree \code{data.frame}
 #'
 #' @examples
 #' ped = data.frame(ID=1:10, SIRE=c(0,0,1,2,0,5,4,4,0,8), DAM=c(0,0,0,3,3,0,6,6,6,0))
@@ -23,7 +23,7 @@ smgsped <- function(ped, pheno) {
   sireofphe = sireofphe[sireofphe > 0]
   dams = unique(ped$DAM)
   dams = dams[dams > 0]
-  # Extract MGS as 
+  # Extract MGS as
   ped = merge(ped, ped[ped$ID %in% dams, c("ID","SIRE")], by.x="DAM", by.y="ID", all.x=TRUE)
   ped = ped[,c("ID","SIRE.x","SIRE.y")]
   colnames(ped) = c("ID","SIRE","MGS")